extract `EnhancedCompositedTransformAnchor`, making `enhanced_composited_transform` completely has no dependency on portal-related things #69
