+++ lib/enhanced_composited_transform.dart
@@ -1,0 +1,6 @@
+export 'src/enhanced_composited_transform/anchor.dart'
+    show EnhancedCompositedTransformAnchor, EnhancedCompositedTransformAligned;
+export 'src/enhanced_composited_transform/flutter_src/widgets_basic.dart'
+    show EnhancedCompositedTransformTarget, EnhancedCompositedTransformFollower;
+export 'src/enhanced_composited_transform/theater_info.dart'
+    show EnhancedCompositedTransformTheaterInfo;
